Правки: Конфиг и вынесение функций из шаблона
