+++ function.php
@@ -6,6 +6,18 @@
  * Time: 0:25
  */
 
+function loading_page($page_loading_function){
+    global $config;
+    if ($config['enable']) {
+
+        $page_loading_function();
+
+    } else {
+        $error_msg = "Сайт на техническом обслуживании";
+        print ($error_msg);
+    }
+};
+
 function render_template($template, $data) {
     $html = '';
     if (file_exists($template)) {
@@ -16,3 +28,13 @@ function render_template($template, $data) {
     }
     return $html;
 };
+
+function modify_price($value) {
+    $changed_price = ceil($value);
+    return $changed_price >= 1000 ? $changed_price = number_format($changed_price, 0, '', ' ') : $changed_price;
+}
+
+function esc($str) {
+    $text = strip_tags($str);
+    return $text;
+}
